Vision Processing Code
This code is designed in two parts. Each part locates a specific
retroreflector. The front panel includes and output that can be used to
test whether or not it reads a vertical or horizontal retroreflector.
Filename is visionProcessing.vi . The name differs from those of the
other vision processing files in that it utilizes camelCaps.